Refactor plot_confusion_matrix_stacked function
This commit refactors the plot_confusion_matrix_stacked function in the utils/plots.py file. It updates the subset labels from 'Train' and 'Validation' to 'Train 5-fold CV' and 'Test' respectively. This change improves the clarity and accuracy of the labels used in the function.

Also, handle missing values in draw_sankey_plot function

This commit modifies the draw_sankey_plot function in the utils/plots.py file to handle missing values in the 'AL Epigenomic Subtype' column. If the column is 'AL Epigenomic Subtype', missing values are filled with 'Not confident'. Otherwise, missing values are filled with 'Unknown'. This change ensures that the function can handle missing values appropriately based on the specific column.

Fix invalid value error in draw_sankey_plot function

This commit fixes an issue in the draw_sankey_plot function in the utils/plots.py file. It corrects the error message for an invalid value of the nan_action parameter. The error message now correctly lists the available options as "drop1", "drop2", "include", or "keep only".